+++ pandoc-lua-engine/src/Text/Pandoc/Lua/Init.hs
@@ -22,7 +22,8 @@ import Control.Monad.Trans (MonadIO (..))
 import Data.Maybe (catMaybes)
 import HsLua as Lua hiding (status, try)
 import HsLua.Core.Run as Lua
-import Text.Pandoc.Class (PandocMonad, readDataFile)
+import Text.Pandoc.Class (PandocMonad)
+import Text.Pandoc.Data (readDataFile)
 import Text.Pandoc.Error (PandocError (PandocLuaError))
 import Text.Pandoc.Lua.Marshal.List (newListMetatable, pushListModule)
 import Text.Pandoc.Lua.PandocLua (PandocLua, liftPandocLua, runPandocLuaWith)
