+++ src/Text/Pandoc/Class.hs
@@ -27,5 +27,5 @@ import Text.Pandoc.Class.CommonState (CommonState (..))
 import Text.Pandoc.Class.PandocMonad
 import Text.Pandoc.Class.PandocIO
 import Text.Pandoc.Class.PandocPure
-import Text.Pandoc.Translations (Translations)
+import Text.Pandoc.Translations.Types (Translations)
 import Text.Pandoc.Class.Sandbox
