+++ src/Text/Pandoc/Readers/LaTeX/Lang.hs
@@ -24,7 +24,8 @@ import Data.Text (Text)
 import qualified Data.Text as T
 import Text.Pandoc.Shared (extractSpaces)
 import Text.Collate.Lang (Lang(..), renderLang)
-import Text.Pandoc.Class (PandocMonad(..), setTranslations)
+import Text.Pandoc.Class (PandocMonad(..))
+import Text.Pandoc.Translations (setTranslations)
 import Text.Pandoc.Readers.LaTeX.Parsing
 import Text.Pandoc.Parsing (updateState, option, getState, QuoteContext(..),
                             withQuoteContext)
